For ANSI ISO C++ compatibility purposes: -long long replaced by Long (typedef long long Long) -array allocated in the stack with a variable size => replace size by MAX_DOMAIN_SIZE in case of compilation problem. -change unassignedVars in tb2solver from Domain type to BTList<Value> type to avoid assert problem when the number of variables is large.

+++ toulbar2/tb2queue.cpp
@@ -6,14 +6,14 @@
 #include "tb2variable.hpp"
 
 
-void Queue::push(DLink<VariableWithTimeStamp> *elt, long long curTimeStamp) {
+void Queue::push(DLink<VariableWithTimeStamp> *elt, Long curTimeStamp) {
     if (elt->content.timeStamp < curTimeStamp) {
         elt->content.timeStamp = curTimeStamp;
         push_back(elt, false);
     }
 }
 
-void Queue::push(DLink<VariableWithTimeStamp> *elt, EventType incdec, long long curTimeStamp) {
+void Queue::push(DLink<VariableWithTimeStamp> *elt, EventType incdec, Long curTimeStamp) {
     elt->content.incdec |= incdec;
     push(elt, curTimeStamp);
 }
